+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## [2.0.6](https://github.com/CloudCannon/bookshop/compare/2.0.5...2.0.6) (2021-10-08)
+
+
+### Bug Fixes
+
+* normalize paths for fast-glob on windows ([d37e9ed](https://github.com/CloudCannon/bookshop/commit/d37e9ed226dda5d32a9105f988a3c013288f494b))
+* use patched node-runner while awaiting next publish ([d98a91b](https://github.com/CloudCannon/bookshop/commit/d98a91b3634c2d29b917fc4e9b26d5d75e54e95f))
+
+
+
 ## [2.0.5](https://github.com/CloudCannon/bookshop/compare/2.0.4...2.0.5) (2021-10-06)